Updates to make clear the requirements for running integration tests

+++ COBie-connector/src/test/integration/COBieBridge.test.ts
@@ -5,7 +5,7 @@
 
 import { expect } from "chai";
 import * as path from "path";
-import { TestUsers, TestUtility } from "@bentley/oidc-signin-tool";
+import { TestUtility } from "@bentley/oidc-signin-tool";
 import { BridgeJobDefArgs, BridgeRunner } from "@bentley/imodel-bridge";
 import { ServerArgs } from "@bentley/imodel-bridge/lib/IModelHubUtils";
 import { ConnectorTestUtils, TestIModelInfo } from "../ConnectorTestUtils";
@@ -14,34 +14,52 @@ import { AccessToken, AuthorizedClientRequestContext } from "@bentley/itwin-clie
 import { BentleyStatus, ClientRequestContext, Logger } from "@bentley/bentleyjs-core";
 import { KnownTestLocations } from "../KnownTestLocations";
 import { HubUtility } from "../HubUtility";
+import { loadTestConfig, TestConfig } from "./TestConfig";
 
 describe("COBie Sample Connector Integration Test (Online)", () => {
 
   let testProjectId: string;
   let requestContext: AuthorizedClientRequestContext;
   let sampleIModel: TestIModelInfo;
+  let testConfig: TestConfig;
 
   before(async () => {
+    testConfig = loadTestConfig();
+
+    ConnectorTestUtils.setupLogging();
+    ConnectorTestUtils.setupDebugLogLevels();
     await ConnectorTestUtils.startBackend();
 
     if (!IModelJsFs.existsSync(KnownTestLocations.outputDir))
       IModelJsFs.mkdirSync(KnownTestLocations.outputDir);
 
     try {
-      requestContext = await TestUtility.getAuthorizedClientRequestContext(TestUsers.regular);
+      const authClient = await TestUtility.getAuthorizationClient(
+        {
+          email: testConfig.TEST_USER_EMAIL,
+          password: testConfig.TEST_USER_PASSWORD
+        },
+        {
+          clientId: testConfig.CLIENT_ID,
+          redirectUri: testConfig.CLIENT_REDIRECT_URI,
+          scope: "openid email profile organization imodelhub context-registry-service:read-only",
+        }
+      );
+      requestContext = new AuthorizedClientRequestContext(await authClient.getAccessToken());
     } catch (error) {
       Logger.logError("Error", `Failed with error: ${error}`);
     }
-    testProjectId = await HubUtility.queryProjectIdByName(requestContext, "imodeljs_sampleConnector_test");
-    const targetIModelId = await HubUtility.recreateIModel(requestContext, testProjectId, "TestSampleConnector");
+
+    const targetIModelId = await HubUtility.recreateIModel(requestContext, testConfig.CONTEXT_ID, "TestSampleConnector");
     expect(undefined !== targetIModelId);
-    sampleIModel = await ConnectorTestUtils.getTestModelInfo(requestContext, testProjectId, "TestSampleConnector");
+    sampleIModel = await ConnectorTestUtils.getTestModelInfo(requestContext, testConfig.CONTEXT_ID, "TestSampleConnector");
   });
 
   after(async () => {
     await ConnectorTestUtils.shutdownBackend();
     IModelJsFs.purgeDirSync(KnownTestLocations.outputDir);
-    IModelJsFs.unlinkSync(path.join(KnownTestLocations.assetsDir, "test.db"));
+    if (IModelJsFs.existsSync(path.join(KnownTestLocations.assetsDir, "test.db")))
+      IModelJsFs.unlinkSync(path.join(KnownTestLocations.assetsDir, "test.db"));
   });
 
   const runConnector = async (bridgeJobDef: BridgeJobDefArgs, serverArgs: ServerArgs, isUpdate: boolean = false, isSchemaUpdate: boolean = false) => {
@@ -59,7 +77,7 @@ describe("COBie Sample Connector Integration Test (Online)", () => {
     const bridgeJobDef = new BridgeJobDefArgs();
     const testSourcePath = path.join(KnownTestLocations.assetsDir, "test.db");
     bridgeJobDef.sourcePath = testSourcePath;
-    bridgeJobDef.bridgeModule = path.join(__dirname, "..", "..", "COBieBridge.js");
+    bridgeJobDef.bridgeModule = path.join(__dirname, "..", "..", "COBieConnector.js");
     const serverArgs = new ServerArgs();
     serverArgs.contextId = testProjectId;
     serverArgs.iModelId = sampleIModel.id;
